Make sure the lock is deleted after a dirty exit.
If the job process is killed with the KILL signal, then the ensure block
will never run, but the monitoring process can still detect the signal
and invoke the on_failure hooks. In the event of a system crash, resque
will invoke the failure hooks for jobs that no longer have a monitoring
process in Worker#prune_dead_workers.